+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## v44.x / v45 (In progress)
+- Music: Some tracks removed.
+
 ## v44 (2019-12-24)
 - Backgrounds: New flag level 2, color-shift backgrounds. Additionally, very eye-straining backgrounds should be less common at all flag levels.
 - Game improvement patches: "Run button" changed to "Run mode"; default run speed and hold Y to slow to a walk.